Remove Markdown processing via Markdig from HtmlXliff
Bloom is using the javascript markdown-it processor for all markdown
conversions, and converting everything to html before calling HtmlXliff.
This is the recommended approach because markdown-it does a better job
with list items that are marked with attributes.

This is one of those rare changes that consist of only feature removal
and code removal.